+++ src/app/shared/interceptors/api-token-interceptor.service.ts
@@ -2,7 +2,7 @@ import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/c
 import { Injectable } from '@angular/core';
 import { firstValueFrom, from, lastValueFrom } from 'rxjs';
 import { environment } from 'src/app/environments/environment';
-import { AuthService } from 'src/app/services/auth.service';
+import { AuthService } from 'src/app/shared/services/auth.service';
 
 @Injectable({
   providedIn: 'root'
